+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 * exam-mode, rollcall-mode を別ファイルで。リファクタ。オーバーホール。
 * パーセンテージグラフを右下に。
 
+## 4.44.1223 (2025-08-14)
+
+- typing.cljs: made `sent?` a member of `app-state`.
+- show-send-reset-display!: check  `(:sent? @app-state)` inside the function. not caller.
+- can not use `delay` on CLJS.
 
 ## 4.44.1218 (2025-08-14)